+++ components/listview/listview.js
@@ -989,7 +989,7 @@ ListView.prototype = {
         e.preventDefault();
         e.stopPropagation();
 
-        self.element.trigger('click', [{elem: item, data: self.settings.dataset[item.attr('aria-posinset')], index: item.index(), originalEvent: e}]);
+        self.element.trigger('click', [{elem: item, data: self.settings.dataset[item.attr('aria-posinset')-1], index: item.index(), originalEvent: e}]);
         return false;
       });
 
@@ -1000,7 +1000,7 @@ ListView.prototype = {
 
         e.preventDefault();
         e.stopPropagation();
-        self.element.trigger('dblclick', [{elem: $(this), data: self.settings.dataset[item.attr('aria-posinset')], index: item.index(), originalEvent: e}]);
+        self.element.trigger('dblclick', [{elem: $(this), data: self.settings.dataset[item.attr('aria-posinset')-1], index: item.index(), originalEvent: e}]);
         return false;
       });
 
@@ -1011,7 +1011,7 @@ ListView.prototype = {
 
         e.preventDefault();
         e.stopPropagation();
-        self.element.trigger('contextmenu', [{elem: $(this), data: self.settings.dataset[item.attr('aria-posinset')], index: item.index(), originalEvent: e}]);
+        self.element.trigger('contextmenu', [{elem: $(this), data: self.settings.dataset[item.attr('aria-posinset')-1], index: item.index(), originalEvent: e}]);
         return false;
       });
     }