+++ components/pager/pager.js
@@ -495,6 +495,10 @@ Pager.prototype = {
       elements = elements.children();
     }
 
+    if (elements.is('table')) {
+      elements = elements.find('tr');
+    }
+
     return elements;
   },